test(refactor): reorganize test file locations
